fix: Make spec uploads persistent
Discourse routinely removes uploads that are not tied to posts anymore. The way we uploaded was not recognized as a file usage, since we used the file URL of the file in the storage instead of the discourse upload URL. That worked, until the clean up kicked in and removed the spec files.

This PR uses the upload URL now and puts in an additional link to the file for good measure.

It also adds a header line, so the post shows up in the category overview more nicely.